Improved drag & drop of cards onto projects. Improved project list. Added support for deleting projects. Changed keyboard shortcuts to use 'ctrl' instead of 'alt'. CSS for adding padding to elements now more flexible. Bugfix: User name not displayed in file list view title.

+++ Artivity.Apid.Modules.Journal/app/partials/directives/art-project-list.js
@@ -6,75 +6,100 @@
             restrict: 'E',
             templateUrl: 'partials/directives/art-project-list.html',
             controller: ProjectListDirectiveController,
-            controllerAs: 't',
-            link: function (scope, element, attr, ctrl) {
-            }
+            controllerAs: 't'
         }
     }
 
     angular.module('explorerApp').controller('ProjectListDirectiveController', ProjectListDirectiveController);
 
-    function ProjectListDirectiveController($rootScope, $scope, $uibModal, projectService, artFileService) {
+    function ProjectListDirectiveController($rootScope, $scope, $uibModal, selectionService, projectService, artFileService) {
         var t = this;
 
-        t.projectList = [];
-
         t.droppedFile = null;
+        t.projectList = [];
         t.addProject = addProject;
-        t.updateProject = updateProject;
-        t.refreshProjectList = refreshProjectList;
         t.addFileToProject = addFileToProject;
-        t.filterBy = filterBy;
-        t.removeFilter = removeFilter;
-
-    
+        t.updateProject = updateProject;
+        t.deleteProject = deleteProject;
+        t.refreshProjectList = refreshProjectList;
+        t.selectedProject = null;
+        t.selectProject = selectProject;
+        t.isDeleting = false;
+        t.toggleDelete = toggleDelete;
 
         initialize();
-        
-        function initialize()
-        {
+
+        function initialize() {
+            // Make the left and right panes resizable.
+            $(".ui-pane-left").resizable({
+                handles: "e"
+            });
+
+            $rootScope.$on('dragStarted', function () {
+                $('.project-item').addClass('drop-target');
+            });
+
+            $rootScope.$on('dragStopped', function () {
+                $('.project-item').removeClass('drop-target');
+            });
+
             t.refreshProjectList();
         }
 
-        function refreshProjectList(){
-            projectService.getAll().then(function(list) {t.projectList = list;});
+        function refreshProjectList() {
+            projectService.getAll().then(function (list) {
+                t.projectList = list;
+
+                $scope.$apply();
+            });
         }
 
-       function updateProject (project) {
-           projectService.selectedProject = project;
+        function updateProject(project) {
+            projectService.selectedProject = project;
+
             var modalInstance = $uibModal.open({
                 animation: true,
                 templateUrl: 'partials/dialogs/update-project-dialog.html',
                 controller: 'UpdateProjectDialogController',
                 controllerAs: 't'
-            });
-
-            modalInstance.result.then(function (account) {
-                console.log("TODO: Reloading Projects");
-
+            }).closed.then(function (account) {
+                t.refreshProjectList();
             });
         };
 
         function addProject() {
-            projectService.create().then(function(result){
-                updateProject(result);
-            })
+            projectService.create().then(function (result) {
+                t.updateProject(result);
+            });
         }
 
-        function addFileToProject(event){
+        function addFileToProject(event) {
             var file = t.droppedFile;
             var project = $(event.target).scope().project;
             // Project is mapped automatically, file manually. this is why the caps of the uri property are different
-            projectService.addFileToProject(project.Uri, file.uri); 
+            projectService.addFileToProject(project.Uri, file.uri);
         }
 
-        function filterBy(project) {
-            artFileService.filterFilesByProject(project.Uri);
+        function deleteProject(project) {
+            projectService.remove(project.Uri).then(function (result) {
+                t.refreshProjectList();
+            });
         }
 
-        function removeFilter() {
-            artFileService.removeFilter();
+        function selectProject(project) {
+            selectionService.selectedItem(project);
+
+            t.selectedProject = project;
+
+            if (project === null) {
+                artFileService.removeFilter();
+            } else {
+                artFileService.filterFilesByProject(project.Uri);
+            }
         }
 
+        function toggleDelete() {
+            t.isDeleting = !t.isDeleting;
+        }
     }
 })();
